fix: restructure the project structure thematically
tests are in the same folder as the components they test
common contains the Header, the SideMenue and the ContentRouter
communication sports the ProxyJSON, APICalls and UserAuth, the last two will be implemented as soon as backendservices are available
input data contains components that allow the user to input data, e.g. DatePicker, InputFields..
show-information consists of the OverView and Review components and has also the OrganisationView which might be not that consistent. opinions?
user-management will contain the possibility to change the password once implemented